refactor(events-page): split to files
PT-1213
