+++ src/domain/events/types.ts
@@ -1,0 +1,1 @@
+export type PlaceOption = { label: string; value: string };
